requirements.txt
annotated-types==0.5.0
anyio==3.7.1
autopep8==2.0.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
dnspython==2.3.0
exceptiongroup==1.2.0
fastapi==0.103.2
Glances==3.4.0.3
gunicorn==21.2.0
h11==0.14.0
httpie==3.2.2
httptools==0.6.0
idna==3.6
importlib-metadata==6.7.0
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
packaging==23.2
protobuf==3.20.3
psutil==5.9.6
pycodestyle==2.10.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pymongo==4.6.0
PySocks==1.7.1
python-dotenv==0.21.1
python-multipart==0.0.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
ujson==5.7.0
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.18.0
zipp==3.15.0
